// be-cloud-ferozi-release/app/src/app-config/models/service.enum.ts
export enum Service {
  VIDEO_ON_DEMAND = 'VIDEO_ON_DEMAND',
  DOUBT_QUERY = 'DOUBT_QUERY',
  APP_SEARCH = 'APP_SEARCH',
  REFERRAL = 'REFERRAL',
  TESTS = 'TESTS',
  DIAGNOSTIC_TESTS = 'DIAGNOSTIC_TESTS',
  PAYMENTS = 'PAYMENTS',
  SPIN_THE_WHEEL = 'SPIN_THE_WHEEL',
  COURSE_JOURNEY = 'COURSE_JOURNEY',
  ANALYTICS_SERVICE = 'ANALYTICS_SERVICE',
  UPGRADE_USER_PRODUCTS = 'UPGRADE_USER_PRODUCTS',
  NOTES = 'NOTES',
  LIVES = 'LIVES',
  PARTIAL_COURSE_JOURNEY = 'PARTIAL_COURSE_JOURNEY',
  AI_DOUBT_QUERY = 'AI_DOUBT_QUERY',
  DOWNLOAD = 'DOWNLOAD',
  ONBOARDING = 'ONBOARDING',
}